+++ src/lol/math/real.h
@@ -1,7 +1,7 @@
 //
 //  Lol Engine
 //
-//  Copyright © 2010—2018 Sam Hocevar <sam@hocevar.net>
+//  Copyright © 2010—2019 Sam Hocevar <sam@hocevar.net>
 //
 //  Lol Engine is free software. It comes without any warranty, to
 //  the extent permitted by applicable law. You can redistribute it
@@ -231,8 +231,8 @@ private:
 public:
     static int DEFAULT_BIGIT_COUNT;
 
-    static inline int bigit_bits() { return 8 * sizeof(bigit_t); }
-    inline int bigit_count() const { return m_mantissa.size(); }
+    static inline int bigit_bits() { return 8 * (int)sizeof(bigit_t); }
+    inline int bigit_count() const { return (int)m_mantissa.size(); }
     inline int total_bits() const { return bigit_count() * bigit_bits(); }
 };
